+++ viktor/__init__.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 
-__version__ = '1.6.8'
-__update_date__ = '2022-04-18_09:39:45'
+__version__ = '1.6.9'
+__update_date__ = '2022-05-14_19:14:01'
 from pathlib import Path
 ROOT_PATH = Path(__file__)